Split Options and Encoding out of regexp/mod.rs
